+++ CotEditor/Sources/CELineNumberView.m
@@ -51,7 +51,8 @@ static const CGFloat  kLineNumberPadding = 3.0;
 
 @implementation CELineNumberView
 
-static const NSString *LineNumberFontName;
+static CGFontRef LineNumberFont;
+static CGFontRef BoldLineNumberFont;
 
 
 #pragma mark Superclass Methods
@@ -65,7 +66,10 @@ static const NSString *LineNumberFontName;
     dispatch_once(&onceToken, ^{
         NSString *defaultFontName = [[NSUserDefaults standardUserDefaults] stringForKey:CEDefaultLineNumFontNameKey];
         NSFont *font = [NSFont fontWithName:defaultFontName size:0] ? : [NSFont paletteFontOfSize:0];
-        LineNumberFontName = [font fontName];
+        NSFont *boldFont = [[NSFontManager sharedFontManager] convertFont:font toHaveTrait:NSBoldFontMask];
+        
+        LineNumberFont = CGFontCreateWithFontName((CFStringRef)[font fontName]);
+        BoldLineNumberFont = CGFontCreateWithFontName((CFStringRef)[boldFont fontName]);
     });
 }
 
@@ -144,15 +148,13 @@ static const NSString *LineNumberFontName;
     // setup font
     CGFloat masterFontSize = [[[self textView] font] pointSize];
     CGFloat fontSize = round(0.9 * masterFontSize);
-    CTFontRef font = CTFontCreateWithName((CFStringRef)LineNumberFontName, fontSize, nil);
+    CTFontRef font = CTFontCreateWithGraphicsFont(LineNumberFont, fontSize, nil, nil);
     
     CGFloat tickLength = ceil(fontSize / 3);
     
-    CGFontRef cgFont = CTFontCopyGraphicsFont(font, NULL);
-    CGContextSetFont(context, cgFont);
+    CGContextSetFont(context, LineNumberFont);
     CGContextSetFontSize(context, fontSize);
     CGContextSetFillColorWithColor(context, [textColor CGColor]);
-    CFRelease(cgFont);
     
     // prepare glyphs
     CGGlyph wrappedMarkGlyph;
@@ -192,10 +194,11 @@ static const NSString *LineNumberFontName;
     
     BOOL isVerticalText = [self orientation] == NSHorizontalRuler;
     NSUInteger tailGlyphIndex = [layoutManager glyphIndexForCharacterAtIndex:[string length]];
+    NSUInteger insertionLocation = [[self textView] selectedRange].location;
     
     // draw line number block
     CGGlyph *digitGlyphsPtr = digitGlyphs;
-    void (^draw_number)(NSUInteger, NSUInteger, CGFloat, BOOL) = ^(NSUInteger lineNumber, NSUInteger lastLineNumber, CGFloat y, BOOL drawsNumber)
+    void (^draw_number)(NSUInteger, NSUInteger, CGFloat, BOOL, BOOL) = ^(NSUInteger lineNumber, NSUInteger lastLineNumber, CGFloat y, BOOL drawsNumber, BOOL isBold)
     {
         if (isVerticalText) {
             // translate y position to horizontal axis
@@ -229,7 +232,16 @@ static const NSString *LineNumberFontName;
                 glyphs[i] = digitGlyphsPtr[numberAt(i, lineNumber)];
             }
             
+            if (isBold) {
+                CGContextSetFont(context, BoldLineNumberFont);
+            }
+            
             CGContextShowGlyphsAtPositions(context, glyphs, positions, digit);  // draw line number
+            
+            if (isBold) {
+                // back to the regular font
+                CGContextSetFont(context, LineNumberFont);
+            }
         }
     };
     
@@ -246,8 +258,9 @@ static const NSString *LineNumberFontName;
     // draw visible line numbers
     for (NSUInteger glyphIndex = visibleGlyphRange.location; glyphIndex < NSMaxRange(visibleGlyphRange); lineNumber++) { // count "real" lines
         NSUInteger charIndex = [layoutManager characterIndexForGlyphAtIndex:glyphIndex];
-        glyphIndex = NSMaxRange([layoutManager glyphRangeForCharacterRange:[string lineRangeForRange:NSMakeRange(charIndex, 0)]
-                                                      actualCharacterRange:NULL]);
+        NSRange lineRange = [string lineRangeForRange:NSMakeRange(charIndex, 0)];
+        glyphIndex = NSMaxRange([layoutManager glyphRangeForCharacterRange:lineRange actualCharacterRange:NULL]);
+        BOOL isSelected = NSLocationInRange(insertionLocation, lineRange);
         
         while (glyphCount < glyphIndex) { // handle wrapped lines
             NSRange range;
@@ -261,8 +274,8 @@ static const NSString *LineNumberFontName;
                 }
                 
             } else {  // new line
-                BOOL drawsNumber = (!isVerticalText || lineNumber % 5 == 0 || lineNumber == 1);
-                draw_number(lineNumber, lastLineNumber, y, drawsNumber);
+                BOOL drawsNumber = (isSelected || !isVerticalText || lineNumber % 5 == 0 || lineNumber == 1);
+                draw_number(lineNumber, lastLineNumber, y, drawsNumber, isSelected);
             }
             
             glyphCount = NSMaxRange(range);
@@ -274,7 +287,7 @@ static const NSString *LineNumberFontName;
                 tailGlyphIndex == glyphIndex &&  // last line
                 ![layoutManager extraLineFragmentTextContainer])  // no extra number
             {
-                draw_number(lineNumber, lastLineNumber, y, YES);
+                draw_number(lineNumber, lastLineNumber, y, YES, isSelected);
             }
             
             lastLineNumber = lineNumber;
@@ -284,9 +297,10 @@ static const NSString *LineNumberFontName;
     // draw the last "extra" line number
     if ([layoutManager extraLineFragmentTextContainer]) {
         NSRect lineRect = [layoutManager extraLineFragmentUsedRect];
+        BOOL isSelected = [string length] == insertionLocation;
         CGFloat y = -NSMinY(lineRect);
         
-        draw_number(lineNumber, lastLineNumber, y, YES);
+        draw_number(lineNumber, lastLineNumber, y, YES, isSelected);
     }
     
     // draw vertical text tics